Add Setup Instructions to README.md
This PR adds instructions that fix an "error" that occurs when you attempt to import the project before running `applyPatches`. 

This is based on info gathered in PR #16, but with added context and recognizable terms for diagnosing the problem. Hopefully, the wording is better in this one and more mergeable. I was having issues getting paperweight setup myself after being told it existed in the PaperMC Discord (I was maintaining a fork of Paper using literally a fork of the Paper repo), but was having issues getting it up and running since this issue isn't super obvious. These instructions should save future folks some wasted diagnostics time, if anything.